Refactor and add static badge image for mobile
