+++ components/Heroes/LeadHero.tsx
