+++ components/Lottie/ExperienceAnimation.tsx
