+++ components/Lottie/ProjectsAnimation.tsx
